geneve: use GRO hint option in the RX path
At the GRO state, when a valid hint option is found, try match the whole
nested headers and try to aggregate on the inner protocol; in case of
mismatch look for matching nested address and port to properly flush on
a per-inner flow basis.

On GRO completion, the (unmodified) nested headers will be considered
part of the (constant) outer geneve encap header so that plain UDP tunnel
segmentation will yield valid wire packets.

In the geneve RX path, when processing a GSO packet carrying a GRO hint
option, update the nested header length fields from the wire packet size
to the GSO-packet one. If the nested header additionally carries a
checksum, convert it to CSUM-partial.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>