geneve: extract hint option at GRO stage
Add helpers for finding a GRO hint option in the geneve header
performing basic sanitization of the option offsets vs the actual
packet layout, validate the option for GRO aggregation and check
the nested header checksum.

The validation helper closely mirrors similar check performed by
the ipv4 and ipv6 gro callbacks, with the additional twist of accessing
the relevant network header via the GRO hint offset.

To validate the nested UDP checksum, leverage the csum completed of
the outer header, similarly to LCO, with the main difference that in
this case we have the outer checksum available.

Use the helpers to extract the hint info at the GRO stage.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>