dev: introduce fixedid features
Some NICs can perform TSO for a certain features set only when the NIC
is allowed to mangle the IP ID; when such feature is not available -
e.g. when the packet is allowed to be fragmented - such drivers have
to strip the relevant features in their own ndo_features_check op.

Introduce support into the kernel core to deal with such constraint, to
pave the ground to deduplicate drivers-specific code.

This also allows any device to expose TSO over tunnel partial capabilities
without any kind of H/W support, as the outer headers for partials
segmentation with static IP ID are constant, and will be leveraged by
the next patches.

The new field is explicitly excluded from the fastpath cachelines, as
it will be touched only for IPv4 TSO packets without the DF bit set.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
---
The relevant drivers that could be simplified are mostly intel ones:
ixgbe, igb, igbvf, igc, but I strongly suspect even txgbe could be
affected.